Better organize player implementation
