+++ Sources/Player/Player+Items.swift
@@ -1,0 +1,194 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import AVFoundation
+import DequeModule
+
+public extension Player {
+    /// The items queued by the player.
+    var items: [PlayerItem] {
+        get {
+            Array(storedItems)
+        }
+        set {
+            let range = storedItems.startIndex..<storedItems.endIndex
+            storedItems.replaceSubrange(range, with: newValue)
+        }
+    }
+
+    /// Items before the current item (not included).
+    /// - Returns: Items.
+    var previousItems: [PlayerItem] {
+        guard let currentIndex else { return [] }
+        return Array(storedItems.prefix(upTo: currentIndex))
+    }
+
+    /// Items past the current item (not included).
+    /// - Returns: Items.
+    var nextItems: [PlayerItem] {
+        guard let currentIndex else { return Array(storedItems) }
+        return Array(storedItems.suffix(from: currentIndex).dropFirst())
+    }
+
+    /// Insert an item before another item. Does nothing if the item already belongs to the deque.
+    /// - Parameters:
+    ///   - item: The item to insert.
+    ///   - beforeItem: The item before which insertion must take place. Pass `nil` to insert the item at the front
+    ///     of the deque.
+    /// - Returns: `true` iff the item could be inserted.
+    @discardableResult
+    func insert(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
+        guard canInsert(item, before: beforeItem) else { return false }
+        if let beforeItem {
+            guard let index = storedItems.firstIndex(of: beforeItem) else { return false }
+            storedItems.insert(item, at: index)
+        }
+        else {
+            storedItems.prepend(item)
+        }
+        return true
+    }
+
+    /// Insert an item after another item. Does nothing if the item already belongs to the deque.
+    /// - Parameters:
+    ///   - item: The item to insert.
+    ///   - afterItem: The item after which insertion must take place. Pass `nil` to insert the item at the back of
+    ///     the deque. If this item does not exist the method does nothing.
+    /// - Returns: `true` iff the item could be inserted.
+    @discardableResult
+    func insert(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
+        guard canInsert(item, after: afterItem) else { return false }
+        if let afterItem {
+            guard let index = storedItems.firstIndex(of: afterItem) else { return false }
+            storedItems.insert(item, at: storedItems.index(after: index))
+        }
+        else {
+            storedItems.append(item)
+        }
+        return true
+    }
+
+    /// Prepend an item to the deque.
+    /// - Parameter item: The item to prepend.
+    /// - Returns: `true` iff the item could be prepended.
+    @discardableResult
+    func prepend(_ item: PlayerItem) -> Bool {
+        insert(item, before: nil)
+    }
+
+    /// Append an item to the deque.
+    /// - Parameter item: The item to append.
+    /// - Returns: `true` iff the item could be appended.
+    @discardableResult
+    func append(_ item: PlayerItem) -> Bool {
+        insert(item, after: nil)
+    }
+
+    /// Move an item before another item.
+    /// - Parameters:
+    ///   - item: The item to move. The method does nothing if the item does not belong to the deque.
+    ///   - beforeItem: The item before which the moved item must be relocated. Pass `nil` to move the item to the
+    ///     front of the deque. If the item does not belong to the deque the method does nothing.
+    /// - Returns: `true` iff the item could be moved.
+    @discardableResult
+    func move(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
+        guard canMove(item, before: beforeItem), let movedIndex = storedItems.firstIndex(of: item) else {
+            return false
+        }
+        if let beforeItem {
+            guard let index = storedItems.firstIndex(of: beforeItem) else { return false }
+            storedItems.move(from: movedIndex, to: index)
+        }
+        else {
+            storedItems.move(from: movedIndex, to: storedItems.startIndex)
+        }
+        return true
+    }
+
+    /// Move an item after another item.
+    /// - Parameters:
+    ///   - item: The item to move.
+    ///   - afterItem: The item after which the moved item must be relocated. Pass `nil` to move the item to the
+    ///     back of the deque. If the item does not belong to the deque the method does nothing.
+    /// - Returns: `true` iff the item could be moved.
+    @discardableResult
+    func move(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
+        guard canMove(item, after: afterItem), let movedIndex = storedItems.firstIndex(of: item) else {
+            return false
+        }
+        if let afterItem {
+            guard let index = storedItems.firstIndex(of: afterItem) else { return false }
+            storedItems.move(from: movedIndex, to: storedItems.index(after: index))
+        }
+        else {
+            storedItems.move(from: movedIndex, to: storedItems.endIndex)
+        }
+        return true
+    }
+
+    /// Remove an item from the deque.
+    /// - Parameter item: The item to remove.
+    func remove(_ item: PlayerItem) {
+        storedItems.removeAll { $0 === item }
+    }
+
+    /// Remove all items in the deque.
+    func removeAllItems() {
+        storedItems.removeAll()
+    }
+}
+
+extension Player {
+    static func items(before index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
+        guard let index else { return [] }
+        let previousIndex = items.index(before: index)
+        guard previousIndex >= 0 else { return [] }
+        return Array(items.suffix(from: previousIndex))
+    }
+
+    static func items(after index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
+        guard let index else { return [] }
+        let nextIndex = items.index(after: index)
+        guard nextIndex < items.count else { return [] }
+        return Array(items.suffix(from: nextIndex))
+    }
+}
+
+private extension Player {
+    func canInsert(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
+        guard let beforeItem else { return true }
+        return storedItems.contains(beforeItem) && !storedItems.contains(item)
+    }
+
+    func canInsert(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
+        guard let afterItem else { return true }
+        return storedItems.contains(afterItem) && !storedItems.contains(item)
+    }
+
+    func canMove(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
+        guard storedItems.contains(item) else { return false }
+        if let beforeItem {
+            guard item !== beforeItem, let index = storedItems.firstIndex(of: beforeItem) else { return false }
+            guard index > 0 else { return true }
+            return storedItems[storedItems.index(before: index)] !== item
+        }
+        else {
+            return storedItems.first !== item
+        }
+    }
+
+    func canMove(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
+        guard storedItems.contains(item) else { return false }
+        if let afterItem {
+            guard item !== afterItem, let index = storedItems.firstIndex(of: afterItem) else { return false }
+            guard index < storedItems.count - 1 else { return true }
+            return storedItems[storedItems.index(after: index)] !== item
+        }
+        else {
+            return storedItems.last !== item
+        }
+    }
+}
