+++ Sources/Player/Player+RemoteCommandCenter.swift
@@ -1,0 +1,106 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import MediaPlayer
+
+extension Player {
+    func updateControlCenter(nowPlayingInfo: NowPlaying.Info) {
+        if !nowPlayingInfo.isEmpty {
+            if nowPlayingSession.nowPlayingInfoCenter.nowPlayingInfo == nil {
+                uninstallRemoteCommands()
+                installRemoteCommands()
+            }
+            nowPlayingSession.nowPlayingInfoCenter.nowPlayingInfo = nowPlayingInfo
+        }
+        else {
+            uninstallRemoteCommands()
+            nowPlayingSession.nowPlayingInfoCenter.nowPlayingInfo = nil
+        }
+    }
+
+    func uninstallRemoteCommands() {
+        commandRegistrations.forEach { registration in
+            nowPlayingSession.remoteCommandCenter.unregister(registration)
+        }
+        commandRegistrations = []
+    }
+
+    private func installRemoteCommands() {
+        commandRegistrations = [
+            playRegistration(),
+            pauseRegistration(),
+            togglePlayPauseRegistration(),
+            previousTrackRegistration(),
+            nextTrackRegistration(),
+            changePlaybackPositionRegistration(),
+            skipBackwardRegistration(),
+            skipForwardRegistration()
+        ]
+    }
+
+    private func playRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.register(command: \.playCommand) { [weak self] _ in
+            self?.play()
+            return .success
+        }
+    }
+
+    private func pauseRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.register(command: \.pauseCommand) { [weak self] _ in
+            self?.pause()
+            return .success
+        }
+    }
+
+    private func togglePlayPauseRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.register(command: \.togglePlayPauseCommand) { [weak self] _ in
+            self?.togglePlayPause()
+            return .success
+        }
+    }
+
+    private func previousTrackRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.previousTrackCommand.isEnabled = false
+        return nowPlayingSession.remoteCommandCenter.register(command: \.previousTrackCommand) { [weak self] _ in
+            self?.returnToPrevious()
+            return .success
+        }
+    }
+
+    private func nextTrackRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.nextTrackCommand.isEnabled = false
+        return nowPlayingSession.remoteCommandCenter.register(command: \.nextTrackCommand) { [weak self] _ in
+            self?.advanceToNext()
+            return .success
+        }
+    }
+
+    private func changePlaybackPositionRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.register(command: \.changePlaybackPositionCommand) { [weak self] event in
+            guard let positionEvent = event as? MPChangePlaybackPositionCommandEvent else { return .commandFailed }
+            self?.seek(near(.init(seconds: positionEvent.positionTime, preferredTimescale: CMTimeScale(NSEC_PER_SEC))))
+            return .success
+        }
+    }
+
+    private func skipBackwardRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.skipBackwardCommand.isEnabled = false
+        nowPlayingSession.remoteCommandCenter.skipBackwardCommand.preferredIntervals = [.init(value: configuration.backwardSkipInterval)]
+        return nowPlayingSession.remoteCommandCenter.register(command: \.skipBackwardCommand) { [weak self] _ in
+            self?.skipBackward()
+            return .success
+        }
+    }
+
+    private func skipForwardRegistration() -> some RemoteCommandRegistrable {
+        nowPlayingSession.remoteCommandCenter.skipForwardCommand.isEnabled = false
+        nowPlayingSession.remoteCommandCenter.skipForwardCommand.preferredIntervals = [.init(value: configuration.forwardSkipInterval)]
+        return nowPlayingSession.remoteCommandCenter.register(command: \.skipForwardCommand) { [weak self] _ in
+            self?.skipForward()
+            return .success
+        }
+    }
+}
