+++ Sources/Player/Player.swift
@@ -47,11 +47,11 @@ public final class Player: ObservableObject, Equatable {
     @Published var _playbackSpeed: PlaybackSpeed = .indefinite
 
     @Published private var currentTracker: CurrentTracker?
-    @Published private var currentItem: CurrentItem = .good(nil)
-    @Published private var storedItems: Deque<PlayerItem>
+    @Published var currentItem: CurrentItem = .good(nil)
+    @Published var storedItems: Deque<PlayerItem>
 
-    // swiftlint:disable:next private_subject
-    var desiredPlaybackSpeedPublisher = PassthroughSubject<Float, Never>()
+    /// The player configuration
+    public let configuration: PlayerConfiguration
 
     /// The type of stream currently played.
     public var streamType: StreamType {
@@ -86,24 +86,18 @@ public final class Player: ObservableObject, Equatable {
     }
 
     /// The current item duration or `.invalid` when not known.
-    private var itemDuration: CMTime {
+    var itemDuration: CMTime {
         queuePlayer.itemDuration
     }
 
     let queuePlayer = QueuePlayer()
-    private let nowPlayingSession: MPNowPlayingSession
+    let nowPlayingSession: MPNowPlayingSession
 
-    public let configuration: PlayerConfiguration
     var cancellables = Set<AnyCancellable>()
+    var commandRegistrations: [any RemoteCommandRegistrable] = []
 
-    private var commandRegistrations: [any RemoteCommandRegistrable] = []
-
-    var backwardSkipTime: CMTime {
-        CMTime(seconds: -configuration.backwardSkipInterval, preferredTimescale: 1)
-    }
-    var forwardSkipTime: CMTime {
-        CMTime(seconds: configuration.forwardSkipInterval, preferredTimescale: 1)
-    }
+    // swiftlint:disable:next private_subject
+    var desiredPlaybackSpeedPublisher = PassthroughSubject<Float, Never>()
 
     /// Create a player with a given item queue.
     /// - Parameters:
@@ -129,7 +123,7 @@ public final class Player: ObservableObject, Equatable {
         configureExternalPlaybackPublisher()
         configurePresentationSizePublisher()
         configureMutedPublisher()
-        configurePlaybackSpeedPublisher()
+        configurePlaybackSpeedPublishers()
 
         configurePlayer()
     }
@@ -159,462 +153,8 @@ public final class Player: ObservableObject, Equatable {
     }
 }
 
-public extension Player {
-    /// Resume playback.
-    func play() {
-        queuePlayer.play()
-    }
-
-    /// Pause playback.
-    func pause() {
-        queuePlayer.pause()
-    }
-
-    /// Toggle playback between play and pause.
-    func togglePlayPause() {
-        if queuePlayer.rate != 0 {
-            queuePlayer.pause()
-        }
-        else {
-            queuePlayer.play()
-        }
-    }
-}
-
-public extension Player {
-    /// Check whether seeking to a specific time is possible.
-    /// - Parameter time: The time to seek to.
-    /// - Returns: `true` if possible.
-    func canSeek(to time: CMTime) -> Bool {
-        guard timeRange.isValidAndNotEmpty else { return false }
-        return timeRange.start <= time && time <= timeRange.end
-    }
-
-    /// Seek to a given position.
-    /// - Parameters:
-    ///   - position: The position to seek to.
-    ///   - smooth: Set to `true` to enable smooth seeking. This allows any currently pending seek to complete before
-    ///     any new seek is performed, preventing unnecessary cancellation. This makes it possible for the playhead
-    ///     position to be moved in a smoother way.
-    ///   - completion: A completion called when seeking ends. The provided Boolean informs
-    ///     whether the seek could finish without being cancelled.
-    func seek(
-        _ position: Position,
-        smooth: Bool = true,
-        completion: @escaping (Bool) -> Void = { _ in }
-    ) {
-        // Mitigates issues arising when seeking to the very end of the range by introducing a small offset.
-        let time = position.time.clamped(to: timeRange, offset: CMTime(value: 1, timescale: 10))
-        guard time.isValid else {
-            completion(true)
-            return
-        }
-        queuePlayer.seek(
-            to: time,
-            toleranceBefore: position.toleranceBefore,
-            toleranceAfter: position.toleranceAfter,
-            smooth: smooth,
-            completionHandler: completion
-        )
-    }
-}
-
-public extension Player {
-    /// Return whether the current player item player can be returned to its default position.
-    /// - Returns: `true` if skipping to the default position is possible.
-    func canSkipToDefault() -> Bool {
-        switch streamType {
-        case .onDemand, .live:
-            return true
-        case .dvr where chunkDuration.isValid:
-            return time < timeRange.end - chunkDuration
-        default:
-            return false
-        }
-    }
-
-    /// Return the current item to its default position.
-    /// - Parameter completion: A completion called when skipping ends. The provided Boolean informs
-    ///   whether the skip could finish without being cancelled.
-    func skipToDefault(completion: @escaping (Bool) -> Void = { _ in }) {
-        switch streamType {
-        case .dvr:
-            seek(after(timeRange.end)) { finished in
-                completion(finished)
-            }
-        default:
-            seek(near(.zero)) { finished in
-                completion(finished)
-            }
-        }
-    }
-}
-
-public extension Player {
-    /// Check whether skipping backward is possible.
-    /// - Returns: `true` if possible.
-    func canSkipBackward() -> Bool {
-        timeRange.isValidAndNotEmpty
-    }
-
-    /// Check whether skipping forward is possible.
-    /// - Returns: `true` if possible.
-    func canSkipForward() -> Bool {
-        guard timeRange.isValidAndNotEmpty else { return false }
-        if itemDuration.isIndefinite {
-            let currentTime = queuePlayer.targetSeekTime ?? time
-            return canSeek(to: currentTime + forwardSkipTime)
-        }
-        else {
-            return true
-        }
-    }
-
-    /// Skip backward.
-    /// - Parameter completion: A completion called when skipping ends. The provided Boolean informs
-    ///   whether the skip could finish without being cancelled.
-    func skipBackward(completion: @escaping (Bool) -> Void = { _ in }) {
-        skip(withInterval: backwardSkipTime, toleranceBefore: .positiveInfinity, toleranceAfter: .zero, completion: completion)
-    }
-
-    /// Skip forward.
-    /// - Parameter completion: A completion called when skipping ends. The provided Boolean informs
-    ///   whether the skip could finish without being cancelled.
-    func skipForward(completion: @escaping (Bool) -> Void = { _ in }) {
-        skip(withInterval: forwardSkipTime, toleranceBefore: .zero, toleranceAfter: .positiveInfinity, completion: completion)
-    }
-
-    private func skip(
-        withInterval interval: CMTime,
-        toleranceBefore: CMTime,
-        toleranceAfter: CMTime,
-        completion: @escaping (Bool) -> Void = { _ in }
-    ) {
-        assert(interval != .zero)
-        let endTolerance = CMTime(value: 1, timescale: 1)
-        let currentTime = queuePlayer.targetSeekTime ?? time
-        if interval < .zero || currentTime < timeRange.end - endTolerance {
-            seek(
-                to(currentTime + interval, toleranceBefore: toleranceBefore, toleranceAfter: toleranceAfter),
-                smooth: true,
-                completion: completion
-            )
-        }
-        else {
-            completion(true)
-        }
-    }
-}
-
-public extension Player {
-    /// Return a publisher periodically emitting the current time while the player is playing content. Does not emit any
-    /// value on subscription and only emits valid times.
-    /// - Parameters:
-    ///   - interval: The interval at which events must be emitted.
-    ///   - queue: The queue on which values are published.
-    /// - Returns: The publisher.
-    func periodicTimePublisher(forInterval interval: CMTime, queue: DispatchQueue = .main) -> AnyPublisher<CMTime, Never> {
-        Publishers.PeriodicTimePublisher(for: queuePlayer, interval: interval, queue: queue)
-    }
-
-    /// Return a publisher emitting when traversing the specified times during normal playback.
-    /// - Parameters:
-    ///   - times: The times to observe.
-    ///   - queue: The queue on which values are published.
-    /// - Returns: The publisher.
-    func boundaryTimePublisher(for times: [CMTime], queue: DispatchQueue = .main) -> AnyPublisher<Void, Never> {
-        Publishers.BoundaryTimePublisher(for: queuePlayer, times: times, queue: queue)
-    }
-}
-
-public extension Player {
-    /// The items queued by the player.
-    var items: [PlayerItem] {
-        get {
-            Array(storedItems)
-        }
-        set {
-            let range = storedItems.startIndex..<storedItems.endIndex
-            storedItems.replaceSubrange(range, with: newValue)
-        }
-    }
-
-    /// Items before the current item (not included).
-    /// - Returns: Items.
-    var previousItems: [PlayerItem] {
-        guard let currentIndex else { return [] }
-        return Array(storedItems.prefix(upTo: currentIndex))
-    }
-
-    /// Return the list of items to be loaded to return to the previous (playable) item.
-    private var returningItems: [PlayerItem] {
-        Self.items(before: currentIndex, in: storedItems)
-    }
-
-    /// Items past the current item (not included).
-    /// - Returns: Items.
-    var nextItems: [PlayerItem] {
-        guard let currentIndex else {
-            return Array(storedItems)
-        }
-        return Array(storedItems.suffix(from: currentIndex).dropFirst())
-    }
-
-    /// Return the list of items to be loaded to advance to the next (playable) item.
-    private var advancingItems: [PlayerItem] {
-        Self.items(after: currentIndex, in: storedItems)
-    }
-
-    private func canInsert(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
-        guard let beforeItem else { return true }
-        return storedItems.contains(beforeItem) && !storedItems.contains(item)
-    }
-
-    /// Insert an item before another item. Does nothing if the item already belongs to the deque.
-    /// - Parameters:
-    ///   - item: The item to insert.
-    ///   - beforeItem: The item before which insertion must take place. Pass `nil` to insert the item at the front
-    ///     of the deque.
-    /// - Returns: `true` iff the item could be inserted.
-    @discardableResult
-    func insert(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
-        guard canInsert(item, before: beforeItem) else { return false }
-        if let beforeItem {
-            guard let index = storedItems.firstIndex(of: beforeItem) else { return false }
-            storedItems.insert(item, at: index)
-        }
-        else {
-            storedItems.prepend(item)
-        }
-        return true
-    }
-
-    private func canInsert(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
-        guard let afterItem else { return true }
-        return storedItems.contains(afterItem) && !storedItems.contains(item)
-    }
-
-    /// Insert an item after another item. Does nothing if the item already belongs to the deque.
-    /// - Parameters:
-    ///   - item: The item to insert.
-    ///   - afterItem: The item after which insertion must take place. Pass `nil` to insert the item at the back of
-    ///     the deque. If this item does not exist the method does nothing.
-    /// - Returns: `true` iff the item could be inserted.
-    @discardableResult
-    func insert(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
-        guard canInsert(item, after: afterItem) else { return false }
-        if let afterItem {
-            guard let index = storedItems.firstIndex(of: afterItem) else { return false }
-            storedItems.insert(item, at: storedItems.index(after: index))
-        }
-        else {
-            storedItems.append(item)
-        }
-        return true
-    }
-
-    /// Prepend an item to the deque.
-    /// - Parameter item: The item to prepend.
-    /// - Returns: `true` iff the item could be prepended.
-    @discardableResult
-    func prepend(_ item: PlayerItem) -> Bool {
-        insert(item, before: nil)
-    }
-
-    /// Append an item to the deque.
-    /// - Parameter item: The item to append.
-    /// - Returns: `true` iff the item could be appended.
-    @discardableResult
-    func append(_ item: PlayerItem) -> Bool {
-        insert(item, after: nil)
-    }
-
-    private func canMove(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
-        guard storedItems.contains(item) else { return false }
-        if let beforeItem {
-            guard item !== beforeItem, let index = storedItems.firstIndex(of: beforeItem) else { return false }
-            guard index > 0 else { return true }
-            return storedItems[storedItems.index(before: index)] !== item
-        }
-        else {
-            return storedItems.first !== item
-        }
-    }
-
-    /// Move an item before another item.
-    /// - Parameters:
-    ///   - item: The item to move. The method does nothing if the item does not belong to the deque.
-    ///   - beforeItem: The item before which the moved item must be relocated. Pass `nil` to move the item to the
-    ///     front of the deque. If the item does not belong to the deque the method does nothing.
-    /// - Returns: `true` iff the item could be moved.
-    @discardableResult
-    func move(_ item: PlayerItem, before beforeItem: PlayerItem?) -> Bool {
-        guard canMove(item, before: beforeItem), let movedIndex = storedItems.firstIndex(of: item) else {
-            return false
-        }
-        if let beforeItem {
-            guard let index = storedItems.firstIndex(of: beforeItem) else { return false }
-            storedItems.move(from: movedIndex, to: index)
-        }
-        else {
-            storedItems.move(from: movedIndex, to: storedItems.startIndex)
-        }
-        return true
-    }
-
-    private func canMove(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
-        guard storedItems.contains(item) else { return false }
-        if let afterItem {
-            guard item !== afterItem, let index = storedItems.firstIndex(of: afterItem) else { return false }
-            guard index < storedItems.count - 1 else { return true }
-            return storedItems[storedItems.index(after: index)] !== item
-        }
-        else {
-            return storedItems.last !== item
-        }
-    }
-
-    /// Move an item after another item.
-    /// - Parameters:
-    ///   - item: The item to move.
-    ///   - afterItem: The item after which the moved item must be relocated. Pass `nil` to move the item to the
-    ///     back of the deque. If the item does not belong to the deque the method does nothing.
-    /// - Returns: `true` iff the item could be moved.
-    @discardableResult
-    func move(_ item: PlayerItem, after afterItem: PlayerItem?) -> Bool {
-        guard canMove(item, after: afterItem), let movedIndex = storedItems.firstIndex(of: item) else {
-            return false
-        }
-        if let afterItem {
-            guard let index = storedItems.firstIndex(of: afterItem) else { return false }
-            storedItems.move(from: movedIndex, to: storedItems.index(after: index))
-        }
-        else {
-            storedItems.move(from: movedIndex, to: storedItems.endIndex)
-        }
-        return true
-    }
-
-    /// Remove an item from the deque.
-    /// - Parameter item: The item to remove.
-    func remove(_ item: PlayerItem) {
-        storedItems.removeAll { $0 === item }
-    }
-
-    /// Remove all items in the deque.
-    func removeAllItems() {
-        storedItems.removeAll()
-    }
-}
-
-public extension Player {
-    internal static let startTimeThreshold: TimeInterval = 3
-
-    private func canReturn(before index: Int?, in items: Deque<PlayerItem>, streamType: StreamType) -> Bool {
-        if configuration.isSmartNavigationEnabled && streamType == .onDemand {
-            return true
-        }
-        else {
-            return Self.canReturnToItem(before: index, in: items)
-        }
-    }
-
-    /// Check whether returning to the previous content is possible.`
-    /// - Returns: `true` if possible.
-    func canReturnToPrevious() -> Bool {
-        canReturn(before: currentIndex, in: storedItems, streamType: streamType)
-    }
-
-    private func isFarFromStartTime() -> Bool {
-        time.isValid && timeRange.isValid && (time - timeRange.start).seconds >= Self.startTimeThreshold
-    }
-
-    private func shouldSeekToStartTime() -> Bool {
-        guard configuration.isSmartNavigationEnabled else { return false }
-        return (streamType == .onDemand && isFarFromStartTime()) || !canReturnToPreviousItem()
-    }
-
-    /// Return to the previous content.
-    func returnToPrevious() {
-        if shouldSeekToStartTime() {
-            seek(near(.zero))
-        }
-        else {
-            returnToPreviousItem()
-        }
-    }
-
-    /// Check whether moving to the next content is possible.`
-    /// - Returns: `true` if possible.
-    func canAdvanceToNext() -> Bool {
-        canAdvanceToNextItem()
-    }
-
-    /// Move to the next content.
-    func advanceToNext() {
-        advanceToNextItem()
-    }
-
-    /// Set the index of the current item.
-    /// - Parameter index: The index to set.
-    func setCurrentIndex(_ index: Int) throws {
-        guard index != currentIndex else { return }
-        guard (0..<storedItems.count).contains(index) else { throw PlaybackError.itemOutOfBounds }
-        let playerItems = AVPlayerItem.playerItems(from: Array(storedItems.suffix(from: index)))
-        queuePlayer.replaceItems(with: playerItems)
-    }
-}
-
 extension Player {
-    private static func items(before index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
-        guard let index else { return [] }
-        let previousIndex = items.index(before: index)
-        guard previousIndex >= 0 else { return [] }
-        return Array(items.suffix(from: previousIndex))
-    }
-
-    private static func items(after index: Int?, in items: Deque<PlayerItem>) -> [PlayerItem] {
-        guard let index else { return [] }
-        let nextIndex = items.index(after: index)
-        guard nextIndex < items.count else { return [] }
-        return Array(items.suffix(from: nextIndex))
-    }
-
-    private static func canReturnToItem(before index: Int?, in items: Deque<PlayerItem>) -> Bool {
-        !Self.items(before: index, in: items).isEmpty
-    }
-
-    private static func canAdvanceToItem(after index: Int?, in items: Deque<PlayerItem>) -> Bool {
-        !Self.items(after: index, in: items).isEmpty
-    }
-
-    /// Check whether returning to the previous item in the deque is possible.`
-    /// - Returns: `true` if possible.
-    func canReturnToPreviousItem() -> Bool {
-        Self.canReturnToItem(before: currentIndex, in: storedItems)
-    }
-
-    /// Return to the previous item in the deque. Skips failed items.
-    func returnToPreviousItem() {
-        guard canReturnToPreviousItem() else { return }
-        queuePlayer.replaceItems(with: AVPlayerItem.playerItems(from: returningItems))
-    }
-
-    /// Check whether moving to the next item in the deque is possible.`
-    /// - Returns: `true` if possible.
-    func canAdvanceToNextItem() -> Bool {
-        Self.canAdvanceToItem(after: currentIndex, in: storedItems)
-    }
-
-    /// Move to the next item in the deque.
-    func advanceToNextItem() {
-        guard canAdvanceToNextItem() else { return }
-        queuePlayer.replaceItems(with: AVPlayerItem.playerItems(from: advancingItems))
-    }
-}
-
-public extension Player {
-    private static func smoothPlayerItem(for currentItem: CurrentItem, in items: Deque<PlayerItem>) -> AVPlayerItem? {
+    static func smoothPlayerItem(for currentItem: CurrentItem, in items: Deque<PlayerItem>) -> AVPlayerItem? {
         switch currentItem {
         case let .bad(playerItem):
             if let lastItem = items.last, lastItem.matches(playerItem) {
@@ -625,40 +165,6 @@ public extension Player {
             }
         case let .good(playerItem):
             return playerItem
-        }
-    }
-
-    /// Check whether the player has finished playing its content and can be restarted.
-    /// - Returns: `true` if possible.
-    func canRestart() -> Bool {
-        guard !storedItems.isEmpty else { return false }
-        return Self.smoothPlayerItem(for: currentItem, in: storedItems) == nil
-    }
-
-    /// Restart playback if possible.
-    func restart() {
-        guard canRestart() else { return }
-        try? setCurrentIndex(0)
-    }
-}
-
-extension Player {
-    private struct Current: Equatable {
-        let item: PlayerItem
-        let index: Int
-    }
-
-    private struct ItemUpdate {
-        let items: Deque<PlayerItem>
-        let currentItem: AVPlayerItem?
-
-        func currentIndex() -> Int? {
-            items.firstIndex { $0.matches(currentItem) }
-        }
-
-        func streamTypePublisher() -> AnyPublisher<StreamType, Never> {
-            guard let currentItem else { return Just(.unknown).eraseToAnyPublisher() }
-            return currentItem.streamTypePublisher().eraseToAnyPublisher()
         }
     }
 
@@ -716,11 +222,11 @@ extension Player {
 
     private func configureControlCenterPublishers() {
         guard !ProcessInfo.processInfo.isiOSAppOnMac else { return }
-        configureControlCenterPublisher()
+        configureControlCenterMetadataPublisher()
         configureControlCenterCommandAvailabilityPublisher()
     }
 
-    private func configureControlCenterPublisher() {
+    private func configureControlCenterMetadataPublisher() {
         Publishers.CombineLatest(
             nowPlayingInfoMetadataPublisher(),
             queuePlayer.nowPlayingInfoPlaybackPublisher()
@@ -751,7 +257,7 @@ extension Player {
                 nowPlayingSession.remoteCommandCenter.skipBackwardCommand.isEnabled = areSkipsEnabled
                 nowPlayingSession.remoteCommandCenter.skipForwardCommand.isEnabled = areSkipsEnabled
                 nowPlayingSession.remoteCommandCenter.previousTrackCommand.isEnabled = canReturn(before: index, in: items, streamType: streamType)
-                nowPlayingSession.remoteCommandCenter.nextTrackCommand.isEnabled = Self.canAdvanceToItem(after: index, in: items)
+                nowPlayingSession.remoteCommandCenter.nextTrackCommand.isEnabled = canAdvanceToItem(after: index, in: items)
             }
             .store(in: &cancellables)
     }
@@ -773,17 +279,6 @@ extension Player {
             .store(in: &cancellables)
     }
 
-    private func assetsPublisher() -> AnyPublisher<[any Assetable], Never> {
-        $storedItems
-            .map { items in
-                Publishers.AccumulateLatestMany(items.map { item in
-                    item.$asset
-                })
-            }
-            .switchToLatest()
-            .eraseToAnyPublisher()
-    }
-
     private func configureExternalPlaybackPublisher() {
         queuePlayer.publisher(for: \.isExternalPlaybackActive)
             .receiveOnMainThread()
@@ -802,139 +297,21 @@ extension Player {
             .assign(to: &$isMuted)
     }
 
-    private func itemUpdatePublisher() -> AnyPublisher<ItemUpdate, Never> {
-        Publishers.CombineLatest($storedItems, $currentItem)
-            .map { items, currentItem in
-                let playerItem = Self.smoothPlayerItem(for: currentItem, in: items)
-                return ItemUpdate(items: items, currentItem: playerItem)
-            }
-            .eraseToAnyPublisher()
-    }
-
-    private func currentPublisher() -> AnyPublisher<Current?, Never> {
-        itemUpdatePublisher()
-            .map { update in
-                guard let currentIndex = update.currentIndex() else { return nil }
-                return .init(item: update.items[currentIndex], index: currentIndex)
+    private func configurePlaybackSpeedPublishers() {
+        playbackSpeedUpdatePublisher()
+            .scan(.indefinite) { speed, update in
+                speed.updated(with: update)
             }
             .removeDuplicates()
-            .eraseToAnyPublisher()
-    }
+            .receiveOnMainThread()
+            .assign(to: &$_playbackSpeed)
 
-    func nowPlayingInfoMetadataPublisher() -> AnyPublisher<NowPlaying.Info, Never> {
-        currentPublisher()
-            .map { current in
-                guard let current else {
-                    return Just(NowPlaying.Info()).eraseToAnyPublisher()
-                }
-                return current.item.$asset
-                    .map { $0.nowPlayingInfo() }
-                    .eraseToAnyPublisher()
+        $_playbackSpeed
+            .sink { [queuePlayer] speed in
+                guard queuePlayer.rate != 0 else { return }
+                queuePlayer.defaultRate = speed.effectiveValue
+                queuePlayer.rate = speed.effectiveValue
             }
-            .switchToLatest()
-            .removeDuplicates { lhs, rhs in
-                // swiftlint:disable:next legacy_objc_type
-                NSDictionary(dictionary: lhs).isEqual(to: rhs)
-            }
-            .eraseToAnyPublisher()
-    }
-}
-
-private extension Player {
-    func playRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.register(command: \.playCommand) { [weak self] _ in
-            self?.play()
-            return .success
-        }
-    }
-
-    func pauseRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.register(command: \.pauseCommand) { [weak self] _ in
-            self?.pause()
-            return .success
-        }
-    }
-
-    func togglePlayPauseRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.register(command: \.togglePlayPauseCommand) { [weak self] _ in
-            self?.togglePlayPause()
-            return .success
-        }
-    }
-
-    func previousTrackRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.previousTrackCommand.isEnabled = false
-        return nowPlayingSession.remoteCommandCenter.register(command: \.previousTrackCommand) { [weak self] _ in
-            self?.returnToPrevious()
-            return .success
-        }
-    }
-
-    func nextTrackRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.nextTrackCommand.isEnabled = false
-        return nowPlayingSession.remoteCommandCenter.register(command: \.nextTrackCommand) { [weak self] _ in
-            self?.advanceToNext()
-            return .success
-        }
-    }
-
-    func changePlaybackPositionRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.register(command: \.changePlaybackPositionCommand) { [weak self] event in
-            guard let positionEvent = event as? MPChangePlaybackPositionCommandEvent else { return .commandFailed }
-            self?.seek(near(.init(seconds: positionEvent.positionTime, preferredTimescale: CMTimeScale(NSEC_PER_SEC))))
-            return .success
-        }
-    }
-
-    func skipBackwardRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.skipBackwardCommand.isEnabled = false
-        nowPlayingSession.remoteCommandCenter.skipBackwardCommand.preferredIntervals = [.init(value: configuration.backwardSkipInterval)]
-        return nowPlayingSession.remoteCommandCenter.register(command: \.skipBackwardCommand) { [weak self] _ in
-            self?.skipBackward()
-            return .success
-        }
-    }
-
-    func skipForwardRegistration() -> some RemoteCommandRegistrable {
-        nowPlayingSession.remoteCommandCenter.skipForwardCommand.isEnabled = false
-        nowPlayingSession.remoteCommandCenter.skipForwardCommand.preferredIntervals = [.init(value: configuration.forwardSkipInterval)]
-        return nowPlayingSession.remoteCommandCenter.register(command: \.skipForwardCommand) { [weak self] _ in
-            self?.skipForward()
-            return .success
-        }
-    }
-
-    func installRemoteCommands() {
-        commandRegistrations = [
-            playRegistration(),
-            pauseRegistration(),
-            togglePlayPauseRegistration(),
-            previousTrackRegistration(),
-            nextTrackRegistration(),
-            changePlaybackPositionRegistration(),
-            skipBackwardRegistration(),
-            skipForwardRegistration()
-        ]
-    }
-
-    func uninstallRemoteCommands() {
-        commandRegistrations.forEach { registration in
-            nowPlayingSession.remoteCommandCenter.unregister(registration)
-        }
-        commandRegistrations = []
-    }
-
-    func updateControlCenter(nowPlayingInfo: NowPlaying.Info) {
-        if !nowPlayingInfo.isEmpty {
-            if nowPlayingSession.nowPlayingInfoCenter.nowPlayingInfo == nil {
-                uninstallRemoteCommands()
-                installRemoteCommands()
-            }
-            nowPlayingSession.nowPlayingInfoCenter.nowPlayingInfo = nowPlayingInfo
-        }
-        else {
-            uninstallRemoteCommands()
-            nowPlayingSession.nowPlayingInfoCenter.nowPlayingInfo = nil
-        }
+            .store(in: &cancellables)
     }
 }
